Moves resfult tests into own file
