+++ src/test/java/tech/sirwellington/alchemy/http/restful/NumValidateTest.kt
@@ -1,0 +1,25 @@
+/*
+ * Copyright © 2018. Sir Wellington.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ *
+ * You may obtain a copy of the License at
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package tech.sirwellington.alchemy.http.restful
+
+import org.junit.runner.RunWith
+import tech.sirwellington.alchemy.annotations.testing.IntegrationTest
+import tech.sirwellington.alchemy.test.junit.runners.AlchemyTestRunner
+
+@RunWith(AlchemyTestRunner::class)
+@IntegrationTest
+class NumValidateTest
